+++ tests/testthat/test-roxytest-testexamples-ASCNworkflow.R
@@ -2,7 +2,7 @@
 
 # File R/ASCNworkflow.R: @testexamples
 
-test_that("Function gcap.ASCNworkflow() @ L59", {
+test_that("Function gcap.ASCNworkflow() @ L70", {
   
   data("ascn")
   data <- ascn
@@ -25,10 +25,21 @@ test_that("Function gcap.ASCNworkflow() @ L59", {
   print(rv$sample_summary)
   print(rv$gene_summary)
   print(rv$cytoband_summary)
+  
+  # Create a subset fCNA
+  rv_subset <- rv$subset(total_cn > 10)
+  nrow(rv$data)
+  nrow(rv_subset$data)
+  
+  rv_subset2 <- rv$subset(sample == "TCGA-02-2485-01")
+  nrow(rv_subset2$data)
+  unique(rv_subset2$data$sample)
   expect_equal(rv, rv2)
   expect_equal(length(rv3), 2L)
   expect_equal(length(rv4), 2L)
   expect_equal(length(rv5), 2L)
   expect_error(gcap.ASCNworkflow(data, outdir = tempdir()))
+  expect_is(rv_subset, "fCNA")
+  expect_is(rv_subset2, "fCNA")
 })